requirements_additions.txt
# SILMARIL Alpha 2.0 — additions to your existing requirements.txt
# Add these lines to your existing requirements.txt (don't replace it).

alpaca-py>=0.20.0
numpy>=1.24
